Struct senna::senna::Senna

pub struct Senna {
    pub senna_ptr: *mut c_void,
    pub pos_map: HashMap<&'static str, POS>,
    pub psg_map: HashMap<&'static str, Phrase>,
}

Handle to all the hashes etc.

Fields
senna_ptr: *mut c_void

pointer to the corresponding C struct

pos_map: HashMap<&'static str, POS>

HashMap for converting string pos tags to pos::POS (unfortunately, there seem to be no static HashMaps in Rust yet)

psg_map: HashMap<&'static str, Phrase>

HashMap for converting string psg tags to phrase::Phrase (unfortunately, there seem to be no static HashMaps in Rust yet)

Methods

impl<'a> Senna

fn new(opt_path: String) -> Senna

Initializes senna based on the data files in opt_path

fn get_number_of_words(&mut self, sentence: &str) -> u32

Returns the number of words contained in a string (warning: Does full tokenization, i.e. relatively costly)

fn parse(&mut self,
sentence: &'a str,
options: SennaParseOptions)
-> Sentence<'a>

Parses a sentence and returns the results as a Sentence structure.