Struct senna::sentence::Word

pub struct Word<'t> { /* fields omitted */ }

Methods

impl<'t> Word<'t>

fn new(offset_start: usize,
offset_end: usize,
string: &'t str,
n: u32)
-> Word<'t>

Constructor for a new word. A POS tag can later be added using set_pos.

fn set_pos(&mut self, tag: POS)

Sets the POS tag

fn get_pos(&self) -> POS

returns the POS tag (empty string if not set)

fn get_string(&self) -> &str

Returns the string representation of the word

fn get_offset_start(&self) -> usize

Returns the offset of the first character of the word in the sentence

fn get_offset_end(&self) -> usize

Returns the offset of the end of the word in the sentence

fn get_n(&self) -> u32

Returns the position of the word in the sentence (n-th word)